hackathon_2022/src/requirements.txt
black~=22.0
flake8>=3.7.9, <4.0
ipython>=7.31.1, <8.0
isort~=5.0
jupyter~=1.0
jupyterlab~=3.0
kedro[spark.SparkDataSet, pandas.ParquetDataSet]==0.18.1
kedro-telemetry~=0.2.0
nbstripout~=0.4
numpy~=1.21
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
pytest~=6.2
snowflake-connector-python==2.7.8
